+++ crates/entrypoint/src/cli/command/terraform/mod.rs
@@ -1,9 +1,9 @@
 pub mod terraform_audit;
 pub mod terraform_command;
+pub mod terraform_reflow;
 pub mod terraform_source;
 pub mod terraform_source_add_imports;
 pub mod terraform_source_generate;
-pub mod terraform_reflow;
 
 use clap::Args;
 use eyre::Result;