+++ crates/hcl/src/discovery/mod.rs
@@ -1,7 +1,7 @@
 mod discover_hcl;
-mod read;
 mod import_block_discoverer;
+mod read;
 
 pub use discover_hcl::*;
-pub use read::*;
 pub use import_block_discoverer::*;
+pub use read::*;