+++ crates/hcl/src/reflow/reflow_by_block_identifier.rs
@@ -14,7 +14,10 @@ impl HclReflower for ReflowByBlockIdentifier {
         let mut reflowed = HashMap::new();
         for (path, body) in hcl {
             let Some(dir) = path.parent() else {
-                bail!("Expected path to have a parent directory: {}", path.display());
+                bail!(
+                    "Expected path to have a parent directory: {}",
+                    path.display()
+                );
             };
             for structure in body.into_iter() {
                 match structure.into_block() {
@@ -33,10 +36,7 @@ impl HclReflower for ReflowByBlockIdentifier {
                                     .map(|attribute| attribute.value.as_str())
                                     .flatten();
                                 if let Some(alias) = alias {
-                                    dir.join(format!(
-                                        "provider.{}.{}.tf",
-                                        provider_name, alias
-                                    ))
+                                    dir.join(format!("provider.{}.{}.tf", provider_name, alias))
                                 } else {
                                     dir.join(format!("provider.{}.tf", provider_name))
                                 }
@@ -49,10 +49,7 @@ impl HclReflower for ReflowByBlockIdentifier {
                                     .unwrap_or("unknown_resource");
                                 let resource_name =
                                     block.labels.get(1).map(|s| s.as_str()).unwrap_or("unnamed");
-                                dir.join(format!(
-                                    "resource.{}.{}.tf",
-                                    resource_type, resource_name
-                                ))
+                                dir.join(format!("resource.{}.{}.tf", resource_type, resource_name))
                             }
                             "data" => {
                                 let data_type = block
@@ -63,6 +60,16 @@ impl HclReflower for ReflowByBlockIdentifier {
                                 let data_name =
                                     block.labels.get(1).map(|s| s.as_str()).unwrap_or("unnamed");
                                 dir.join(format!("data.{}.{}.tf", data_type, data_name))
+                            }
+                            "variable" => {
+                                let variable_name =
+                                    block.labels.get(0).map(|s| s.as_str()).unwrap_or("unnamed");
+                                dir.join(format!("variable.{}.tf", variable_name))
+                            }
+                            "output" => {
+                                let output_name =
+                                    block.labels.get(0).map(|s| s.as_str()).unwrap_or("unnamed");
+                                dir.join(format!("output.{}.tf", output_name))
                             }
                             _ => path.clone(),
                         };