+++ crates/hcl/src/reflow/reflow_trait.rs
@@ -1,12 +1,9 @@
 use hcl::edit::structure::Body;
 use std::collections::HashMap;
-
 use std::path::PathBuf;
 
 #[async_trait::async_trait]
 pub trait HclReflower: Send {
-    async fn reflow(
-        &mut self,
-        hcl: HashMap<PathBuf, Body>,
-    ) -> eyre::Result<HashMap<PathBuf, Body>>;
+    async fn reflow(&mut self, hcl: HashMap<PathBuf, Body>)
+    -> eyre::Result<HashMap<PathBuf, Body>>;
 }